moved ByteArrayMarshaller to LogHelperTest
